Clone works as expected now, and join works basically

+++ kernel/trap.c
@@ -84,11 +84,12 @@ trap(struct trapframe *tf)
       panic("trap");
     }
     // In user space, assume process misbehaved.
-    cprintf("pid %d %s: trap %d err %d on cpu %d "
-            "eip 0x%x addr 0x%x--kill proc\n",
-            proc->pid, proc->name, tf->trapno, tf->err, cpu->id, tf->eip, 
-            rcr2());
-    proc->killed = 1;
+	if(rcr2() != 0xffffffff)
+		cprintf("pid %d %s: trap %d err %d on cpu %d "
+				"eip 0x%x addr 0x%x--kill proc\n",
+				proc->pid, proc->name, tf->trapno, tf->err, cpu->id, tf->eip, 
+				rcr2());
+	proc->killed = 1;
   }
 
   // Force process exit if it has been killed and is in user space.
